requirements/production.txt
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
camelot-py==0.7.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
distro==1.4.0
entrypoints==0.3
et-xmlfile==1.0.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.2
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.3
opencv-python==4.1.1.26
openpyxl==3.0.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pdfminer.six==20191110
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
pycodestyle==2.5.0
pycryptodome==3.9.3
Pygments==2.4.2
pyparsing==2.4.5
PyPDF2==1.26.0
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
pywin32==226
pywinpty==0.5.5
pyzmq==18.1.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.13.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.10
tabula-py==1.4.1
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
xlwt==1.3.0
zipp==0.6.0
